Preserve user rotation on lockscreen when auto rotate disabled
By default, when auto-rotate is disabled, the lockscreen ignores the
last user-selected orientation and falls back to portrait. This is
problematic tor tablet users, if the device was last used in landscape,
the lockscreen still appear in portrait after the screen turns off
and back on.

This change updates DisplayRotation to restore the user’s locked
rotation (mUserRotation) whenever the keyguard is showing and
auto-rotate is disabled. As a result, the lockscreen now remembers the
last orientation across screen off/on, without affecting normal app rotation.

Change-Id: Ieb66e25bbcafb1fc61088bf993d0a0a36754c4f8